backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pytest
httpx

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
slowapi==0.1.9

# Google OAuth
google-auth
google-auth-oauthlib

# Google Gemini API
google-generativeai